feat: support template users
